Addition of comment regarding delay in measurement loop
User romankaelin asked whether delay in measurement loop was necessary. I'd seen this done in a few arduino sketches but on closer examination I don't think it is necessary here. Really the delay was serving the purpose of slowing the update rate of the LCD screen. I think a better way to implement this would be to have a timer that causes the LCD to report at a fixed rate. It's something to add in a future update.