Update to add non blocking timer
This update is designed to remove the need for a blocking delay in the timer. The ReportDelay variable holds the minimum time (500ms) between lcd refreshes. When the current time minus the last reporting time exceeds this delay lcdReport is called. This function clears the lcd, and prints the last updated values for voltage and steinhart. Credit to user romankaelin for highlighting the limitations of using a blocking delay in the previous iteration of the code.